+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- added `CacheProvider<TEnum>` utility class
+- export the `StandardEnum<T>` type
+
 ## [0.3.1] - 2024-01-19
 
 - added export `lib/string` and `lib/guid`